backend/requirements.txt
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.19
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.2
scikit-learn==1.6.1
aif360==0.6.1
fairlearn==0.11.0
shap==0.46.0
lime==0.2.0.1
reportlab==4.2.5
python-dotenv==1.0.1
pydantic==2.10.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
joblib==1.4.2
matplotlib==3.9.4
